requirements.txt
aniso8601==1.3.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
grpcio==1.7.0
grpcio-tools==1.7.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
peewee==2.10.2
protobuf==3.4.0
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
sqlite-web==0.1.8
urllib3==1.22
Werkzeug==0.12.2
